update: changed id param type to long in intersection repository file

+++ src/main/java/dev/rygen/intersectionlightcontroller/repositories/IntersectionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IntersectionRepository extends JpaRepository<Intersection, Integer> {
+public interface IntersectionRepository extends JpaRepository<Intersection, Long> {
 }
